extras/lambdas/transcribe_SP/requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.23.2
assemblyai==0.20.2
requests==2.31.0
python-multipart==0.0.6
guardrails-ai==0.3.2
mangum
pydantic
validators